Add a bunch of CGFloat casts

+++ components/TextFields/examples/experimental/supplemental/MDCFilledTextField.m
@@ -119,11 +119,11 @@
 @implementation MDCFilledTextFieldPositioningDelegate
 
 - (CGFloat)assistiveLabelPaddingWithContainerHeight:(CGFloat)containerHeight {
-  return 0.13 * containerHeight;
+  return (CGFloat)0.13 * containerHeight;
 }
 
 - (CGFloat)defaultContainerHeightWithTextHeight:(CGFloat)textHeight {
-  return 3.3 * textHeight;
+  return (CGFloat)3.3 * textHeight;
 }
 
 - (CGFloat)containerHeightWithTextHeight:(CGFloat)textHeight
@@ -139,8 +139,8 @@
                   preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.28;
-  return offset - (0.5 * floatingLabelHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.28;
+  return offset - ((CGFloat)0.5 * floatingLabelHeight);
 }
 
 - (CGFloat)textMinYWithFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -148,8 +148,8 @@
                           preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.64;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.64;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 - (CGFloat)textMinYWithoutFloatingLabelWithTextHeight:(CGFloat)textHeight
@@ -157,8 +157,8 @@
                              preferredContainerHeight:(CGFloat)preferredContainerHeight {
   CGFloat containerHeight = [self containerHeightWithTextHeight:textHeight
                                        preferredContainerHeight:preferredContainerHeight];
-  CGFloat offset = containerHeight * 0.5;
-  return offset - (0.5 * textHeight);
+  CGFloat offset = containerHeight * (CGFloat)0.5;
+  return offset - ((CGFloat)0.5 * textHeight);
 }
 
 @end
